requirements.txt
mistralai
python-dotenv
pypdf
boto3